ENG-5298: Fix double deletion when both elastic context and COW context are in use.
If a tuple is pending delete by the COW context, the elatic context shouldn't
try to delete it again. Let the COW context take care of it after it's safely
scanned.

+++ src/ee/storage/ElasticIndexReadContext.cpp
@@ -199,7 +199,9 @@ void ElasticIndexReadContext::deleteStreamedTuples()
     m_iter->reset();
     TableTuple tuple;
     while (m_iter->next(tuple)) {
-        m_surgeon.deleteTuple(tuple);
+        if (!tuple.isPendingDelete()) {
+            m_surgeon.deleteTuple(tuple);
+        }
     }
 }
